feat(multiple): Combined L1 syncing stages, l1Syncer cache, tests
