+++ turbo/jsonrpc/zkevm_api_test.go
@@ -5,6 +5,7 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"github.com/erigontech/erigon-lib/kv/memdb"
 	"math"
 	"math/big"
 	"testing"
@@ -373,6 +374,8 @@ func TestGetBatchByNumber(t *testing.T) {
 
 	db := contractBackend.DB()
 	agg := contractBackend.Agg()
+	_, cacheDB := context.Background(), memdb.NewTestDB(t)
+	defer cacheDB.Close()
 
 	baseApi := NewBaseApi(nil, stateCache, contractBackend.BlockReader(), agg, false, rpccfg.DefaultEvmCallTimeout, contractBackend.Engine(), datadir.New(t.TempDir()))
 	ethImpl := NewEthAPI(baseApi, db, nil, nil, nil, 5000000, 100_000, 100_000, &ethconfig.Defaults, false, 100, 100, log.New(), defaultL1GasPriceTracker, 1000, false)
@@ -383,6 +386,7 @@ func TestGetBatchByNumber(t *testing.T) {
 
 	l1Syncer := syncer.NewL1Syncer(
 		ctx,
+		cacheDB,
 		[]syncer.IEtherman{EthermanMock},
 		[]common.Address{},
 		[][]common.Hash{},
@@ -653,6 +657,9 @@ func TestGetExitRootsByGER(t *testing.T) {
 	db := contractBackend.DB()
 	agg := contractBackend.Agg()
 
+	_, cacheDB := context.Background(), memdb.NewTestDB(t)
+	defer cacheDB.Close()
+
 	baseApi := NewBaseApi(nil, stateCache, contractBackend.BlockReader(), agg, false, rpccfg.DefaultEvmCallTimeout, contractBackend.Engine(), datadir.New(t.TempDir()))
 	ethImpl := NewEthAPI(baseApi, db, nil, nil, nil, 5000000, 100_000, 100_000, &ethconfig.Defaults, false, 100, 100, log.New(), defaultL1GasPriceTracker, 1000, false)
 
@@ -662,6 +669,7 @@ func TestGetExitRootsByGER(t *testing.T) {
 
 	l1Syncer := syncer.NewL1Syncer(
 		ctx,
+		cacheDB,
 		[]syncer.IEtherman{EthermanMock},
 		[]common.Address{},
 		[][]common.Hash{},
@@ -728,6 +736,9 @@ func TestLatestGlobalExitRoot(t *testing.T) {
 	db := contractBackend.DB()
 	agg := contractBackend.Agg()
 
+	_, cacheDB := context.Background(), memdb.NewTestDB(t)
+	defer cacheDB.Close()
+
 	baseApi := NewBaseApi(nil, stateCache, contractBackend.BlockReader(), agg, false, rpccfg.DefaultEvmCallTimeout, contractBackend.Engine(), datadir.New(t.TempDir()))
 	ethImpl := NewEthAPI(baseApi, db, nil, nil, nil, 5000000, 100_000, 100_000, &ethconfig.Defaults, false, 100, 100, log.New(), defaultL1GasPriceTracker, 1000, false)
 
@@ -737,6 +748,7 @@ func TestLatestGlobalExitRoot(t *testing.T) {
 
 	l1Syncer := syncer.NewL1Syncer(
 		ctx,
+		cacheDB,
 		[]syncer.IEtherman{EthermanMock},
 		[]common.Address{},
 		[][]common.Hash{},
